# api-docs/schemas/user-service-openapi.yaml
openapi: 3.1.0
info:
  title: User Management Service API
  description: REST API for user authentication and authorization
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: API Gateway
security:
- Bearer Authentication: []
tags:
- name: Authentication
  description: User registration and authentication endpoints
paths:
  /api/v1/auth/register:
    post:
      tags:
      - Authentication
      summary: Register a new user
      description: "Creates a new user account with ROLE_USER by default. Password\
        \ must contain at least 8 characters, one digit, one uppercase letter, one\
        \ lowercase letter, and one special character."
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "409":
          description: Email already exists
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "400":
          description: Invalid input data
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "201":
          description: User registered successfully
          content:
            '*/*':
              schema:
                type: string
  /api/v1/auth/login:
    post:
      tags:
      - Authentication
      summary: Login
      description: "Authenticates user and returns JWT access token. Token must be\
        \ included in subsequent requests as: Authorization: Bearer {token}"
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Invalid credentials or account disabled
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: "Login successful, JWT token returned"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LoginResponse"
  /api/v1/auth/change-password:
    post:
      tags:
      - Authentication
      summary: Change password
      description: Changes the password for the currently authenticated user. Requires
        current password for verification. New password must meet security requirements.
      operationId: changePassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Unauthorized - invalid current password or token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "400":
          description: Invalid password or new password same as current
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: Password changed successfully
          content:
            '*/*':
              schema:
                type: string
      security:
      - Bearer Authentication: []
  /api/internal/auth/validate:
    post:
      tags:
      - internal-auth-controller
      operationId: validateToken
      parameters:
      - name: Authorization
        in: header
        required: true
        schema:
          type: string
      responses:
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserValidationResponse"
  /api/v1/auth/me:
    get:
      tags:
      - Authentication
      summary: Get current user information
      description: "Returns information about the currently authenticated user including\
        \ id, email, roles, and account status"
      operationId: getCurrentUser
      responses:
        "409":
          description: Conflict
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Unauthorized - invalid or missing token
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "200":
          description: User information retrieved successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserInfoResponse"
      security:
      - Bearer Authentication: []
components:
  schemas:
    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        properties:
          type: object
          additionalProperties: {}
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          minLength: 0
        password:
          type: string
          maxLength: 20
          minLength: 8
          pattern: "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%^&+=!]).*$"
      required:
      - email
      - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
      - email
      - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        type:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          maxLength: 20
          minLength: 8
          pattern: "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[@#$%^&+=!]).*$"
      required:
      - currentPassword
      - newPassword
    UserValidationResponse:
      type: object
      properties:
        valid:
          type: boolean
        username:
          type: string
        roles:
          type: array
          items:
            type: string
    UserInfoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        roles:
          type: array
          items:
            type: string
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
  securitySchemes:
    Bearer Authentication:
      type: http
      description: Enter JWT token
      scheme: bearer
      bearerFormat: JWT